+++ packages/mindmap/src/constants/index.ts
@@ -11,8 +11,8 @@ export const STROKE_WIDTH = 3;
 export const EXTEND_OFFSET = 8;
 export const EXTEND_RADIUS = 16;
 
-/* 快捷插入圆直径 */
-export const QUICK_INSERT_CIRCLE_DIAMETER = 18;
 export const QUICK_INSERT_CIRCLE_OFFSET = 9;
+export const QUICK_INSERT_CIRCLE_COLOR = '#6698FF';
+export const QUICK_INSERT_INNER_CROSS_COLOR = 'white';
 
 export * from './node';